Starting GameplayTag Refactor
Major changes on this one. Tore down a bunch of the code and rebuilt it to run off of GameplayTags. This commit is in the early stages of the transition so there's a pretty big mash up of code in there. I expect a lot of the older stuff to get wiped out and totally rebuilt.

Added a bunch of input files to manage movement and interaction. Added a chest to test that interaction. Will re work door and button and finish up the elevator on the next commit.