Portal Update
Have portal system working and operating. Still needs tuning and performance updates but you can throw portals and it will teleport you to the other one and the view is correct. Teleport Location needs to be dialed in as well as the orientation and some visual weirdness when placing on opposing walls.